backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
PyYAML==6.0.2
kubernetes==33.1.0
python-dotenv==1.1.1